% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPrice.R
\name{getPrice}
\alias{getPrice}
\title{Get current coin price from Coingecko}
\usage{
getPrice(coin_id, currency)
}
\arguments{
\item{coin_id}{The id of the coin on Coingecko}

\item{currency}{The currency used to display the coin price}
}
\description{
This function gets the current price of a given token from Coingecko
}
